feature userAccess : tests

+++ packages/types-database/src/types/provider/userAccess/userIpMask.ts
@@ -1,4 +1,5 @@
 import { Schema } from "mongoose";
+import type {Decimal128} from "mongodb";
 
 // ip mask is usually present like this: '198.51.100.14/24'
 // but since IP has its numeric presentation, and the mask is a rigid range,
@@ -6,8 +7,8 @@ import { Schema } from "mongoose";
 
 interface UserIpMask {
 	// ipV6 takes 128bits, so we can't use Number, which is only 54 bits
-	rangeMin: bigint;
-	rangeMax: bigint;
+	rangeMin: Decimal128;
+	rangeMax: Decimal128;
 	// CIDR prefix https://en.wikipedia.org/wiki/Classless_Inter-Domain_Routing - 198.51.100.14/{24}
 	// for presentation only purposes
 	numeric: number;
